docs: Navigation 컴포넌트 폴더 구조 수정
- 불필요하게 나눠져 있는 컴포넌트 결합
- 폴더 구조 수정으로 import 경로 수정
- 폴더 구조 변경에 따른 test 코드 수정
