--- fe/src/pages/Layout.tsx
@@ -1,13 +1,0 @@
-import { Outlet } from "react-router-dom";
-import Navigation from "@/components/Navigation/Navigation";
-
-const Layout = () => {
-    return (
-        <>
-            <Navigation/>
-            <Outlet />
-        </>
-    );
-};
-
-export default Layout;
