+++ fe/src/tests/pagesTests/Layout.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
-import Layout from "@/pages/Layout";
+import DefaultLayout from "@/components/Layout/DefaultLayout";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 const queryClient = new QueryClient();
@@ -10,7 +10,7 @@ test("Navigation이 렌더링 되었을때 Outlet이 렌더링 되는지 확인"
         <MemoryRouter initialEntries={["/test"]}>
             <QueryClientProvider client={queryClient}>
                 <Routes>
-                    <Route path="/" element={<Layout />}>
+                    <Route path="/" element={<DefaultLayout />}>
                         <Route
                             path="test"
                             element={<div>Test Component</div>}
